move NpFillNan and FillNanMethod to fina_utils module
